requirements.txt
requests==2.31.0
PyQt5==5.15.9
openai==0.27.0
python-dotenv==0.19.2